# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for user subscriptions.",
        "title": "Swagger Subscription Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/subscription": {
            "put": {
                "description": "Update subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Update subscription",
                "parameters": [
                    {
                        "description": "update subscription",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "strig"
                        }
                    }
                }
            },
            "post": {
                "description": "Create subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create subscription",
                "parameters": [
                    {
                        "description": "create subscription",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "strig"
                        }
                    }
                }
            }
        },
        "/subscription/{id}": {
            "get": {
                "description": "Get subscriptions by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscription by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subscription Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete subscriptions by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Delete subscription by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subscription Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "description": "Get add subscriptions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get add subscriptions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Subscription"
                            }
                        }
                    }
                }
            }
        },
        "/subscriptions/total_summ": {
            "get": {
                "description": "Get summ subscriptions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get summ subscriptions",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "service name",
                        "name": "service_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "01-2006",
                        "description": "From date",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "01-2006",
                        "description": "To date",
                        "name": "to_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateSubscriptionRequest": {
            "type": "object",
            "required": [
                "price",
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "default": "07-2026"
                },
                "price": {
                    "type": "integer",
                    "default": 400
                },
                "service_name": {
                    "type": "string",
                    "default": "Yandex Plus",
                    "maxLength": 150,
                    "minLength": 3
                },
                "start_date": {
                    "type": "string",
                    "default": "07-2025"
                },
                "user_id": {
                    "type": "string",
                    "default": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "model.Subscription": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.UpdateSubscriptionRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}